requirements.txt
Flask==2.3.3
pandas==2.2.2
google-api-python-client==2.128.0
python-dotenv==1.0.1

